Create atm.vhd
At first, we have declared an entity ATM with Key as Input and O as Output.Key is the
User’s Choice i.e. user will have 4 choices. Then, we have declared Architecture. Because
of Input Limitations, we have predefined the Account Number as AN, the Account Balance
as AB, X the Amount to be Deposited/Withdrawn, PIN1 as the ATM PIN of the User and
PIN as the NEW PIN of the User.
User will have 4 choices:-
1)If User Presses Key = 1 then User will be able to Deposit the money.
2)If User Presses Key = 2 then User will be able to Withdraw the money.
3)If User Presses Key = 3 then User will be able to Check the Account Balance.
4)If User Presses Key = 4 then User will be able to Change the PIN.
If money are deposited, then Output is Account Balance. If money to be withdrawn i.e. X
are greater than the Account Balance, then Output is the Account Balance. If X is less,
then Output is X. If PIN is changed, then Output is the New PIN.